+++ orca-api/src/main/java/com/netflix/spinnaker/orca/api/pipeline/ExecutionPreprocessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Netflix, Inc.
+ * Copyright 2021 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.orca.extensionpoint.pipeline;
+package com.netflix.spinnaker.orca.api.pipeline;
 
+import com.netflix.spinnaker.kork.annotations.Beta;
+import com.netflix.spinnaker.kork.plugins.api.internal.SpinnakerExtensionPoint;
 import java.util.Map;
 import javax.annotation.Nonnull;
 
-/** A preprocessor that can modify an Execution upon initial receipt of the configuration. */
-public interface ExecutionPreprocessor {
+/**
+ * ExecutionPreprocessor is a hook point that can modify an Execution upon initial receipt of the
+ * configuration.
+ *
+ * <p>A common use case is to inspect and update a pipeline with configured context, such an
+ * application name.
+ */
+@Beta
+public interface ExecutionPreprocessor extends SpinnakerExtensionPoint {
 
   /** Returns whether or not the preprocess can handle the inbound execution. */
   boolean supports(@Nonnull Map<String, Object> execution, @Nonnull Type type);
